backend/requirements.txt
# ============================================================================
# IMPORTANT: PyTorch must be installed FIRST (before other requirements)
# ============================================================================
# For GPU support (NVIDIA CUDA - RECOMMENDED for 10-50x speedup):
#   pip install torch torchvision torchaudio --index-url https://download.pytorch.org/whl/cu124
#
# For CPU-only (slower, but works without GPU):
#   pip install torch torchvision torchaudio --index-url https://download.pytorch.org/whl/cpu
#
# After installing PyTorch, run: pip install -r requirements.txt
# ============================================================================

# Core FastAPI dependencies
fastapi==0.115.6
uvicorn[standard]==0.34.0
pydantic==2.10.6
python-multipart==0.0.20

# Database and storage
supabase==2.10.0
python-dotenv==1.0.0

# HTTP client
httpx==0.27.2

# Video processing dependencies (for chunked upload system)
opencv-python==4.12.0.88
supervision==0.26.1
ultralytics==8.3.217

# AI/ML dependencies for video analysis
# Note: Using latest versions compatible with Python 3.13
numpy>=1.26.0
umap-learn>=0.5.5
scikit-learn>=1.4.0  # Required for K-means clustering in team assignment
transformers>=4.40.0
sentencepiece>=0.2.0
protobuf>=4.25.0
numba>=0.59.0
pandas>=2.0.0  # Required for ball position interpolation

# Additional utilities
requests==2.31.0

# Letta AI Integration - Intelligent sports coaching
letta-client
tqdm>=4.66.0  # Progress bars for video processing

# Reka-AI
reka-api

# Fish Audio SDK
fish-audio-sdk
pyaudio